+++ packages/cinterop/src/commonMain/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -130,7 +130,7 @@ expect object RealmInterop {
     fun <T> realm_results_max(results: NativePointer, property: Long): T
     fun <T> realm_results_min(results: NativePointer, property: Long): T
     // FIXME OPTIMIZE Get many
-    fun <T> realm_results_get(results: NativePointer, index: Long): T
+    fun realm_results_get(results: NativePointer, index: Long): Link
 
     fun realm_get_object(realm: NativePointer, link: Link): NativePointer