+++ packages/library-base/src/commonMain/kotlin/io/realm/query/RealmQuery.kt
@@ -202,12 +202,6 @@ interface RealmElementQuery<T : RealmObject> : Flowable<RealmResults<T>> {
 
 /**
  * Query returning a single [RealmObject].
- *
- * TODO : The interaction with primitive queries might be a bit awkward
- * TODO : answer to above from C-API:
- *  Note: This function can only produce objects, not values. Use the
- *       `realm_results_t` returned by `realm_query_find_all()` to retrieve
- *        values from a list of primitive values.
  */
 interface RealmSingleQuery<T> : Flowable<T> {